% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(n = 5, p = 0.5)
}
\arguments{
\item{n}{trials}

\item{p}{probability of success}
}
\value{
returns dataframe with amount of success for each number of trials
}
\description{
gives a dataframe with successes and probabilities
}
