% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(n = 10, p = 0.3)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
the kurtosis (taildness) of a distribution as a numeric value
}
\description{
gives the kurtosis for a given distribution
}
