% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(n = 10, p = 0.3)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
the mean number of successes we would expect
}
\description{
gives the mean expectation of a distribution
}
