% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(n = 10, p = 0.3)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
the skewness of a distribution as a given value
}
\description{
gives the skewness of a given distribution
}
