% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(n = 10, p = 0.3)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
the variance of a distribution
}
\description{
gives the varies for a given distribution
}
